Refactor to new wai_runner backend
