+++ .github/workflows/build-win-mac.yml
@@ -13,9 +13,9 @@ jobs:
           python-version: '3.11'
       - run: python -m pip install --upgrade pip
       - run: pip install -r python/runner/requirements.txt pyinstaller
-      - run: pyinstaller python/runner/kestrel_runner.py --onefile --name kestrel_runner
+      - run: pyinstaller python/runner/wai_runner.py --onefile --name wai_runner
       - run: mkdir plugin/WildlifeAI.lrplugin/bin/win
-      - run: copy dist\kestrel_runner.exe plugin\WildlifeAI.lrplugin\bin\win\
+      - run: copy dist\wai_runner.exe plugin\WildlifeAI.lrplugin\bin\win\
       - run: python scripts/package_plugin.py
       - uses: actions/upload-artifact@v4
         with:
@@ -31,9 +31,9 @@ jobs:
           python-version: '3.11'
       - run: python -m pip install --upgrade pip
       - run: pip install -r python/runner/requirements.txt pyinstaller
-      - run: pyinstaller python/runner/kestrel_runner.py --onefile --name kestrel_runner
+      - run: pyinstaller python/runner/wai_runner.py --onefile --name wai_runner
       - run: mkdir -p plugin/WildlifeAI.lrplugin/bin/mac
-      - run: cp dist/kestrel_runner plugin/WildlifeAI.lrplugin/bin/mac/
+      - run: cp dist/wai_runner plugin/WildlifeAI.lrplugin/bin/mac/
       - run: python scripts/package_plugin.py
       - uses: actions/upload-artifact@v4
         with:
