+++ .github/workflows/build.yml
@@ -13,9 +13,9 @@ jobs:
       - run: |
           python -m pip install --upgrade pip
           pip install -r python/runner/requirements.txt pyinstaller
-          pyinstaller python/runner/wai_runner.py --onefile --name kestrel_runner
+          pyinstaller python/runner/wai_runner.py --onefile --name wai_runner
           mkdir plugin/WildlifeAI.lrplugin/bin/win
-          copy dist\kestrel_runner.exe plugin/WildlifeAI.lrplugin\bin\win\
+          copy dist\wai_runner.exe plugin/WildlifeAI.lrplugin\bin\win\
           python scripts/package_plugin.py
       - uses: actions/upload-artifact@v4
         with:
