+++ plugin/WildlifeAI.lrplugin/PluginInit.lua
@@ -3,8 +3,8 @@ local LrPathUtils = import 'LrPathUtils'
 local Log = dofile( LrPathUtils.child(_PLUGIN.path, 'utils/Log.lua') )
 
 local prefs = LrPrefs.prefsForPlugin()
-prefs.runnerWin       = prefs.runnerWin       or 'bin/win/kestrel_runner.exe'
-prefs.runnerMac       = prefs.runnerMac       or 'bin/mac/kestrel_runner'
+prefs.runnerWin       = prefs.runnerWin       or 'bin/win/wai_runner.exe'
+prefs.runnerMac       = prefs.runnerMac       or 'bin/mac/wai_runner'
 prefs.keywordRoot     = prefs.keywordRoot     or 'WildlifeAI'
 if prefs.enableLogging == nil then prefs.enableLogging = false end
 prefs.enableStacking  = prefs.enableStacking  or false
