+++ plugin/WildlifeAI.lrplugin/bin/mac/README.txt
@@ -1,1 +1,1 @@
-Place kestrel_runner (built from python/runner/wai_runner.py) here
+Place wai_runner (built from python/runner/wai_runner.py) here
