+++ scripts/freeze_mac.sh
@@ -4,7 +4,7 @@ python3 -m venv venv
 source venv/bin/activate
 pip install --upgrade pip
 pip install -r python/runner/requirements.txt pyinstaller
-pyinstaller python/runner/wai_runner.py --onefile --name kestrel_runner
+pyinstaller python/runner/wai_runner.py --onefile --name wai_runner
 mkdir -p plugin/WildlifeAI.lrplugin/bin/mac
-cp dist/kestrel_runner plugin/WildlifeAI.lrplugin/bin/mac/
+cp dist/wai_runner plugin/WildlifeAI.lrplugin/bin/mac/
 python scripts/package_plugin.py
