+++ scripts/freeze_win.bat
@@ -4,7 +4,7 @@ python -m venv venv
 call venv\Scripts\activate
 pip install --upgrade pip
 pip install -r python\runner\requirements.txt pyinstaller
-pyinstaller python\runner\wai_runner.py --onefile --name kestrel_runner
-copy dist\kestrel_runner.exe plugin\WildlifeAI.lrplugin\bin\win\
+pyinstaller python\runner\wai_runner.py --onefile --name wai_runner
+copy dist\wai_runner.exe plugin\WildlifeAI.lrplugin\bin\win\
 python scripts\package_plugin.py
 endlocal
